+++ tests/run-all-tests.ts
@@ -30,6 +30,7 @@ await runImport('./unit/complex/literal-enum.test.ts');
 
 // Unit: special
 await runImport('./unit/special/uri-url-buffer-jsonschema.test.ts');
+await runImport('./unit/special/refine.test.ts');
 
 // Behavior
 await runImport('./behavior/type-inference.test.ts');